pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sgcc-electricity-feishu"
dynamic = ["version"]
description = "国家电网飞书应用开发"
readme = "README.md"
requires-python = "<3.13"
license = "MIT"
keywords = []
authors = [
  { name = "lanseria", email = "your-email@example.com" },
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
]
dependencies = [
  "typer>=0.9.0",
  "rich>=13.0.0",
  "python-dotenv>=1.1.0",
  "selenium>=4.31.0",
  "requests>=2.32.3",
  "schedule>=1.2.2",
  "Pillow>=11.1.0",
  "onnxruntime>=1.21.0",
  "numpy>=2.2.4",
  "lark-oapi>=1.4.13",
]

[project.optional-dependencies]
test = [
  "pytest>=7.0.0",
]
build = [
  "pyinstaller>=6.3.0",
]

[project.scripts]
sef = "sgcc_electricity_feishu.main:app"

# Hatchling 特定配置
[tool.hatch.version]
path = "src/sgcc_electricity_feishu/__init__.py"

# 添加 hatch 包配置
[tool.hatch.packages.find]
include = ["src/*"]
src = "src"

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = ["tests"]
python_files = ["test_*.py"]
